Jump to v1 (new API, README, extra functionality) (#10)
This updates the API substantially. The public functions are now:
- `isvd`
- `ISVD.update!`        (not exported)
- `ISVD.impute_nans!`   (not exported)
- `ISVD.Cache`          (not exported)

Caching is new, a way to reduce memory usage. `isvd` is new, a wrapper around `update!`.
And now this can update with a chunk possessing fewer columns than the number of
components, allowing one to use the full data matrix.

Finally, a real README and more method docstrings were added.

Jumps to v1.0.0.